restructure guides, improve audio inferencing performance, and define fixes for common issues.
The initial release of this document provided some basic code snippets and walkthroughs, but did not address a number of complexities when using ccstudio with Edge Impulse. This updates to address many observed issues, including stack and heap configuration, syscfg changes, workarounds to platform and SimpleLink SDK issues. These updates necessitated changes to the structure of the guides.